앱 분할, 모델 수정
