+++ back/accounts/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from potato.models import User
+# Register your models here.
+
+admin.site.register(User)
