+++ back/back/settings.py
@@ -39,6 +39,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'potato',
+    'back',
+    'accounts',
     'corsheaders',
 ]
 
@@ -78,23 +80,23 @@ WSGI_APPLICATION = 'back.wsgi.application'
 # Database
 # https://docs.djangoproject.com/en/4.2/ref/settings/#databases
 
-DATABASES = {
-   'default': {
-       'ENGINE': 'django.db.backends.sqlite3',
-       'NAME': BASE_DIR / 'db.sqlite3',
-   }
-}
-
 # DATABASES = {
-#     "default": {
-#         "ENGINE": "django.db.backends.mysql",
-#         "NAME": "potato",  # DB name
-#         "USER": "root",  # DB account
-#         "PASSWORD": "Rokmc1218!",  # DB account's password
-#         "HOST": "localhost",  # DB address(IP)
-#         "PORT": "3306",  # DB port(normally 3306)
-#     }
+#    'default': {
+#        'ENGINE': 'django.db.backends.sqlite3',
+#        'NAME': BASE_DIR / 'db.sqlite3',
+#    }
 # }
+
+DATABASES = {
+    "default": {
+        "ENGINE": "django.db.backends.mysql",
+        "NAME": "potato",  # DB name
+        "USER": "root",  # DB account
+        "PASSWORD": "Rokmc1218!",  # DB account's password
+        "HOST": "localhost",  # DB address(IP)
+        "PORT": "3306",  # DB port(normally 3306)
+    }
+}
 
 # Password validation
 # https://docs.djangoproject.com/en/4.2/ref/settings/#auth-password-validators
