+++ back/back/urls.py
@@ -16,6 +16,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path,include
+from ninja import NinjaAPI
 # from .api import api
 from django.conf import settings
 from django.conf.urls.static import static
