+++ back/potato/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Comment,User,Timer
+from .models import Comment,StudyTimer
 # Register your models here.
 admin.site.register(Comment)
-admin.site.register(User)
-admin.site.register(Timer)
+admin.site.register(StudyTimer)
