+++ back/requirements.txt
@@ -1,12 +1,47 @@
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
 asgiref==3.7.2
+asttokens==2.2.1
+backcall==0.2.0
+bcrypt==4.0.1
+certifi==2023.7.22
+cffi==1.15.1
+charset-normalizer==3.2.0
+colorama==0.4.6
+decorator==5.1.1
 distlib==0.3.6
 Django==3.2.13
+django-appconf==1.0.5
+django-cleanup==8.0.0
+django-cors-headers==4.2.0
+django-extensions==3.2.3
+django-imagekit==4.1.0
 django-ninja==0.22.2
+executing==1.2.0
 filelock==3.12.2
+idna==3.4
+ipython==8.14.0
+jedi==0.18.2
+matplotlib-inline==0.1.6
 mysqlclient==2.2.0
-Pillow==10.0.1
+parso==0.8.3
+pickleshare==0.7.5
+pilkit==2.0
+Pillow==9.5.0
 platformdirs==3.7.0
+prompt-toolkit==3.0.38
+pure-eval==0.2.2
+pycparser==2.21
 pydantic==1.10.12
+Pygments==2.15.1
 pytz==2023.3.post1
+requests==2.31.0
+six==1.16.0
 sqlparse==0.4.4
+stack-data==0.6.2
+traitlets==5.9.0
 typing_extensions==4.7.1
+tzdata==2023.3
+urllib3==2.0.4
+virtualenv==20.23.1
+wcwidth==0.2.6
