+++ back/schedule/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from potato.models import Comment
+# Register your models here.
+
+admin.site.register(Comment)
